* AudaciousUDPStub/AudaciousUDPStub.c: Implemented PositionUpdate and   TitleUpdate messages. * OpenVP/UDPPlayerData.cs: Changed PlayerData.SongPosition from double   to float.  Removed some debug code. * OpenVP/PlayerData.cs: Changed PlayerData.SongPosition from double to   float. * AudaciousUDPStub/ChangeLog, ChangeLog, OpenVP.Core/ChangeLog,   OpenVP/ChangeLog: Added.

+++ OpenVP/PlayerData.cs
@@ -175,7 +175,7 @@ namespace OpenVP {
 		/// <value>
 		/// The current position of the song in fractional seconds.
 		/// </value>
-		public abstract double SongPosition { get; }
+		public abstract float SongPosition { get; }
 		
 		/// <summary>
 		/// The current song title.
